+++ examples/tutorial/test_tutorials.py
@@ -17,6 +17,7 @@ import pytest
                              'step5',
                              'step6',
                              'step7',
+                             'step8',
                              'reactivex']
 
                          )