+++ rsocket/handlers/interfaces.py
@@ -1,0 +1,2 @@
+class Requester:
+    pass