+++ rsocket/handlers/request_channel_requester.py
@@ -4,19 +4,20 @@ from typing import Optional
 from reactivestreams.publisher import Publisher
 from reactivestreams.subscriber import Subscriber
 from rsocket.frame_builders import to_request_channel_frame
+from rsocket.handlers.interfaces import Requester
 from rsocket.handlers.request_cahnnel_common import RequestChannelCommon
 from rsocket.payload import Payload
 from rsocket.rsocket import RSocket
 
 
-class RequestChannelRequester(RequestChannelCommon):
+class RequestChannelRequester(RequestChannelCommon, Requester):
 
     def __init__(self,
                  socket: RSocket,
                  payload: Payload,
-                 remote_publisher: Optional[Publisher] = None,
+                 publisher: Optional[Publisher] = None,
                  sending_done: Optional[asyncio.Event] = None):
-        super().__init__(socket, remote_publisher, sending_done)
+        super().__init__(socket, publisher, sending_done)
         self._payload = payload
 
     def setup(self):
@@ -27,7 +28,7 @@ class RequestChannelRequester(RequestChannelCommon):
             to_request_channel_frame(stream_id=self.stream_id,
                                      payload=payload,
                                      initial_request_n=self._initial_request_n,
-                                     complete=self._remote_publisher is None,
+                                     complete=self._publisher is None,
                                      fragment_size_bytes=self.socket.get_fragment_size_bytes())
         )
 
@@ -36,5 +37,5 @@ class RequestChannelRequester(RequestChannelCommon):
         super().subscribe(subscriber)
         self._send_channel_request(self._payload)
 
-        if self._remote_publisher is None:
+        if self._publisher is None:
             self.mark_completed_and_finish(sent=True)