+++ rsocket/handlers/request_response_requester.py
@@ -2,6 +2,7 @@ import asyncio
 
 from rsocket.frame import ErrorFrame, PayloadFrame, Frame, error_frame_to_exception
 from rsocket.frame_builders import to_request_response_frame
+from rsocket.handlers.interfaces import Requester
 from rsocket.helpers import create_future, payload_from_frame
 from rsocket.local_typing import Awaitable
 from rsocket.payload import Payload
@@ -9,7 +10,7 @@ from rsocket.rsocket import RSocket
 from rsocket.streams.stream_handler import StreamHandler
 
 
-class RequestResponseRequester(StreamHandler):
+class RequestResponseRequester(StreamHandler, Requester):
     def __init__(self, socket: RSocket, payload: Payload):
         super().__init__(socket)
         self._payload = payload