+++ tests/rsocket/helpers.py
@@ -1,7 +1,9 @@
 import asyncio
 import json
+import os
+from contextlib import contextmanager
 from dataclasses import dataclass
-from datetime import timedelta
+from datetime import timedelta, datetime
 from math import ceil
 from typing import Tuple, Any
 from typing import Type, Callable
@@ -102,3 +104,28 @@ def to_json_bytes(item: Any) -> bytes:
 
 def create_data(base: bytes, multiplier: int, limit: float = None):
     return b''.join([ensure_bytes(str(i)) + base for i in range(multiplier)])[0:limit]
+
+
+def create_large_random_data(size: int):
+    return bytearray(os.urandom(size))
+
+
+def benchmark_method(method: Callable, iterations: int) -> float:
+    with measure_runtime() as result:
+        for i in range(iterations):
+            method()
+
+    return result.time.total_seconds() * 1000 / iterations
+
+
+@dataclass
+class Result:
+    time: timedelta = None
+
+
+@contextmanager
+def measure_runtime():
+    result = Result()
+    start = datetime.now()
+    yield result
+    result.time = datetime.now() - start